requirements.txt
# Core Dependencies
langchain==0.1.10
langchain-community>=0.0.25
langchain-openai==0.0.8
langchain-qdrant==0.1.3
qdrant-client>=1.10.1
openai>=1.3.0

# Confluence Integration
atlassian-python-api==3.41.0

# Environment Variable Management
python-dotenv==1.0.0

# Optional: For better logging and debugging
python-json-logger==2.0.7
